Change license from GPL-2 to LPGL-3
This should have about the same implications for contributors but make
it easier for SE to sanction and embed in sites by default if they ever
choose to do so. It should also be more conducive to other projects that
want to re-use parts of the code for a different purpose.