Fix Makefile bugs where prefixes ignored
The @ and - prefixes are only being looked at on the first lines of each
recipe because of .ONESHELL. See documentation at:

http://www.gnu.org/software/make/manual/make.html#One-Shell

This fixes the behaviour to be more as expected since half my prefixes
weren't doing what  they were supposed to and cleans up the output. Less
is more here because of the way console output is buffered in parallel
builds.