Account for more of Travis's make, git and jq being stupid
So in order to reference commits in the parent repo, we actually need a
link to the repo. With Travis's git being too dumb for worktree, we're
doing a separate clone. This links it back up with the master so all the
commit sha's will be available for reference.

Also account for .ONESHELL not actually working in Travis.